added Remove Beer function for Admin
