+++ client/store/beer.js
@@ -4,10 +4,10 @@ import history from '../history'
 /**
  * ACTION TYPES
  */
-const GET_beers = 'GET_beers'
-const ADD_beer = 'ADD_beer'
-const EDIT_beer = 'EDIT_beer'
-const REMOVE_beer = 'REMOVE_beer'
+const GET_BEERS = 'GET_BEERS'
+const ADD_BEER = 'ADD_BEER'
+const EDIT_BEER = 'EDIT_BEER'
+const REMOVE_BEER = 'REMOVE_BEER'
 
 /**
  * INITIAL STATE
@@ -17,10 +17,10 @@ const defaultbeers = []
 /**
  * ACTION CREATORS
  */
-const _getbeers = beers => ({type: GET_beers, beers})
-const _addbeer = beer => ({type: ADD_beer, beer})
-const _editBeer = beer => ({type: EDIT_beer, beer})
-const _removeBeer = beer => ({type: REMOVE_beer, beer})
+const _getbeers = beers => ({type: GET_BEERS, beers})
+const _addBeer = beer => ({type: ADD_BEER, beer})
+const _editBeer = beer => ({type: EDIT_BEER, beer})
+const _removeBeer = beer => ({type: REMOVE_BEER, beer})
 
 /**
  * THUNK CREATORS
@@ -34,11 +34,10 @@ export const getBeers = () => async dispatch => {
   }
 }
 
-export const addbeer = () => async dispatch => {
+export const addBeer = id => async dispatch => {
   // try {
-  //   const beers = (await axios.get('/api/beers')).data
-  //   console.log(beers)
-  //   dispatch(_getbeers(beers))
+  //   const beers = (await axios.post('/api/beers/:id')).data
+  //   dispatch(_addBeer(beer))
   // } catch (err) {
   //   console.error(err)
   // }
@@ -47,17 +46,17 @@ export const editBeer = beer => async dispatch => {
   try {
     const updatedBeer = (await axios.put(`/api/beers/${beer.productId}`, beer))
       .data
-    console.log(updatedBeer, 'updated beer')
     dispatch(_editBeer(updatedBeer))
   } catch (err) {
     console.error(err)
   }
 }
 
-export const removeBeer = id => async dispatch => {
+export const removeBeer = beerId => async dispatch => {
   try {
-    await axios.delete(`/api/beeers/${id}`) // deletes from backend
-    dispatch(_removeBeer(id)) // delete from frontend
+    console.log(beerId)
+    await axios.delete(`/api/beers/${beerId}`)
+    dispatch(_removeBeer(beerId))
   } catch (err) {
     console.error(err)
   }
@@ -68,18 +67,19 @@ export const removeBeer = id => async dispatch => {
  */
 export default function(state = defaultbeers, action) {
   switch (action.type) {
-    case GET_beers:
+    case GET_BEERS:
       return action.beers
-    case ADD_beer:
+    case ADD_BEER:
       return [...state, action.beer]
-    case EDIT_beer:
+    case EDIT_BEER:
       return state.map(beer => {
         if (beer.id === action.beer.id) {
           return action.beer
         } else return beer
       })
-    case REMOVE_beer:
-      return state.filter(beerItem => beerItem.id !== action.beer.id)
+    case REMOVE_BEER:
+      console.log(action.beer, 'in reducer remove')
+      return state.filter(beerItem => beerItem.id !== action.beer)
     default:
       return state
   }
