+++ server/api/beers.js
@@ -2,6 +2,15 @@ const router = require('express').Router()
 const {Product} = require('../db/models')
 module.exports = router
 
+const isAdmin = (req, res, next) => {
+  if (!req.user || !req.user.level) {
+    const err = new Error(`Admin level only!`)
+    err.status = 401
+    return next(err)
+  }
+  next()
+}
+
 router.get('/', async (req, res, next) => {
   try {
     const beers = await Product.findAll({where: {category: 'beer'}})
@@ -13,8 +22,8 @@ router.get('/', async (req, res, next) => {
 
 router.post('/', async (req, res, next) => {
   try {
-    // const beers = await Wine.findAll()
-    // const beers = await Product.findAll({where:{category:'wine'}})
+    // const beers = await Beer.findAll()
+    // const beers = await Product.findAll({where:{category:'beer'}})
     // res.json(beers)
   } catch (err) {
     next(err)
@@ -34,7 +43,6 @@ router.put('/:id', async (req, res, next) => {
         plain: true
       }
     )
-    console.log(product[1], 'product in put')
     res.json(product[1])
   } catch (err) {
     next(err)
@@ -43,7 +51,9 @@ router.put('/:id', async (req, res, next) => {
 
 router.delete('/:id', async (req, res, next) => {
   try {
+    console.log(req.params.id, 'req params')
     const beer = await Product.findByPk(req.params.id)
+    console.log(beer, 'in route')
     await beer.destroy()
     res.sendStatus(204)
   } catch (err) {
